+++ evalml/tuners/tuner_exceptions.py
@@ -1,0 +1,2 @@
+class NoParamsException(Exception):
+    pass