+++ docs/source/changelog.rst
@@ -7,6 +7,7 @@ Changelog
         * Add normalization option and information to confusion matrix :pr:`484`
         * Renamed `PipelineBase.name` as `PipelineBase.summary` and redefined `PipelineBase.name` as class property :pr:`491`
         * Added access to parameters in Pipelines with `PipelineBase.parameters` (used to be return of `PipelineBase.describe`) :pr:`501`
+        * Added `fill_value` parameter for SimpleImputer :pr:`509`    
         * Added functionality to override component hyperparemeters and made pipelines take hyperparemeters from components :pr:`516`
     * Fixes
     * Changes
@@ -14,6 +15,7 @@ Changelog
         * Support pandas 1.0.0 :pr:`486`
         * Made all references to the logger static :pr:`503`
         * Refactored `model_type` parameter for components and pipelines to `model_family` :pr:`507`
+        * Moved `pipelines/utils.save_pipeline` and `pipelines/utils.load_pipeline` to `PipelineBase.save` and `PipelineBase.load` :pr:`526`
     * Documentation Changes
         * Updated API reference to remove PipelinePlot and added moved PipelineBase plotting methods :pr:`483`
         * Add code style and github issue guides :pr:`463` :pr:`512`
@@ -29,6 +31,7 @@ Changelog
     * Components and Pipelines now have a `model_family` field instead of `model_type`
     * `get_pipelines` utility function now accepts `model_families` as an argument instead of `model_types`
     * `PipelineBase.name` no longer returns structure of pipeline and has been replaced by `PipelineBase.summary`
+    * `pipelines/utils.save_pipeline` and `pipelines/utils.load_pipeline` moved to `PipelineBase.save` and `PipelineBase.load`
     
 
 **v0.7.0 Mar. 9, 2020**
@@ -55,6 +58,7 @@ Changelog
         * Dropped support for Python 3.5 :pr:`438`
         * Removed unused `apply.py` file :pr:`449`
         * Clean up requirements.txt to remove unused deps :pr:`451`
+        * Support installation without all required dependencies :pr:`459`
     * Documentation Changes
         * Update release.md with instructions to release to internal license key :pr:`354`
     * Testing Changes
@@ -89,6 +93,7 @@ Changelog
         * Standarizing inputs as pd.Dataframe / pd.Series :pr:`130`
         * Enforcing that pipelines must have an estimator as last component :pr:`277`
         * Added ipywidgets as a dependency in requirements.txt :pr:`278`
+        * Added Random and Grid Search Tuners :pr:`240`
     * Documentation Changes
         * Adding class properties to API reference :pr:`244`
         * Fix and filter FutureWarnings from scikit-learn :pr:`249`, :pr:`257`